Revert "fix: load disruptions from s3 for dashboard (#402)"
This reverts commit 79935d1666fbbbc183186f3ea4bb47ad23d94ddb.
