+++ site/pages/api/get-all-disruptions/get-all-disruptions.test.ts
@@ -1,15 +1,12 @@
-import { PublishStatus } from "@create-disruptions-data/shared-ts/enums";
+import { MiscellaneousReason, PublishStatus, Severity, VehicleMode } from "@create-disruptions-data/shared-ts/enums";
+import * as sharedUtils from "@create-disruptions-data/shared-ts/utils";
 import MockDate from "mockdate";
 import { describe, it, expect, afterEach, beforeEach, vi, afterAll } from "vitest";
-import getAllDisruptions, { getDisruptionsForTable } from "./[organisationId].api";
-import * as s3 from "../../../data/s3";
-import {
-    DEFAULT_ORG_ID,
-    disruptionArray,
-    disruptionWithConsequencesAndSocialMediaPosts,
-    getMockRequestAndResponse,
-    mockSession,
-} from "../../../testData/mockData";
+import getAllDisruptions, { formatSortedDisruption } from "./[organisationId].api";
+import * as dynamo from "../../../data/dynamo";
+import { FullDisruption } from "../../../schemas/disruption.schema";
+import { DEFAULT_ORG_ID, getMockRequestAndResponse, mockSession, sortedDisruption } from "../../../testData/mockData";
+import * as utils from "../../../utils";
 import * as session from "../../../utils/apiUtils/auth";
 
 describe("getAllDisruptions", () => {
@@ -19,10 +16,15 @@ describe("getAllDisruptions", () => {
         getDisruptionsDataFromDynamo: vi.fn(),
     }));
 
-    vi.mock("../../../data/s3");
+    vi.mock("../../utils", async () => ({
+        ...(await vi.importActual<object>("../../utils")),
+        filterDisruptionsForOperatorUser: vi.fn(),
+    }));
 
-    const getDisruptionsSpy = vi.spyOn(s3, "getObject");
+    const getDisruptionsDataFromDynamoSpy = vi.spyOn(dynamo, "getDisruptionsDataFromDynamo");
+    const sortDisruptionsByStartDateSpy = vi.spyOn(sharedUtils, "sortDisruptionsByStartDate");
     const getSessionSpy = vi.spyOn(session, "getSession");
+    const filterDisruptionsForOperatorUserSpy = vi.spyOn(utils, "filterDisruptionsForOperatorUser");
 
     beforeEach(() => {
         getSessionSpy.mockImplementation(() => {
@@ -38,8 +40,75 @@ describe("getAllDisruptions", () => {
         MockDate.reset();
     });
 
+    const disruptions: FullDisruption[] = [
+        {
+            associatedLink: "",
+            consequences: [
+                {
+                    consequenceIndex: 0,
+                    consequenceType: "networkWide",
+                    description: "testpending81",
+                    disruptionDelay: "",
+                    disruptionId: "00831a54-0ecb-4b0c-8d73-310888968747",
+                    disruptionSeverity: Severity.unknown,
+                    removeFromJourneyPlanners: "no",
+                    vehicleMode: VehicleMode.bus,
+                },
+            ],
+            description: "testpending81",
+            displayId: "415351",
+            disruptionId: "00831a54-0ecb-4b0c-8d73-310888968747",
+            disruptionNoEndDateTime: "true",
+            disruptionReason: MiscellaneousReason.accident,
+            disruptionRepeats: "doesntRepeat",
+            disruptionRepeatsEndDate: "",
+            disruptionStartDate: "07/03/2023",
+            disruptionStartTime: "1000",
+            disruptionType: "planned",
+            publishStartDate: "07/03/2023",
+            publishStartTime: "1000",
+            publishStatus: PublishStatus.draft,
+            socialMediaPosts: [],
+            summary: "testpending81",
+            validity: [],
+            template: false,
+        },
+        {
+            associatedLink: "",
+            consequences: [
+                {
+                    consequenceIndex: 0,
+                    consequenceType: "networkWide",
+                    description: "testfilter3",
+                    disruptionDelay: "",
+                    disruptionId: "01b15519-41b5-4ace-a212-5331a1622771",
+                    disruptionSeverity: Severity.unknown,
+                    removeFromJourneyPlanners: "no",
+                    vehicleMode: VehicleMode.bus,
+                },
+            ],
+            description: "testfilter3",
+            displayId: "124b84",
+            disruptionId: "01b15519-41b5-4ace-a212-5331a1622771",
+            disruptionNoEndDateTime: "true",
+            disruptionReason: MiscellaneousReason.accident,
+            disruptionRepeats: "doesntRepeat",
+            disruptionRepeatsEndDate: "",
+            disruptionStartDate: "08/03/2023",
+            disruptionStartTime: "1000",
+            disruptionType: "planned",
+            publishStartDate: "07/03/2023",
+            publishStartTime: "1000",
+            publishStatus: PublishStatus.published,
+            socialMediaPosts: [],
+            summary: "testfilter3",
+            validity: [],
+            template: false,
+        },
+    ];
+
     it("should be successful when session is set", async () => {
-        getDisruptionsSpy.mockResolvedValue("");
+        getDisruptionsDataFromDynamoSpy.mockResolvedValue({ disruptions });
 
         const { req, res } = getMockRequestAndResponse({
             query: {
@@ -54,12 +123,13 @@ describe("getAllDisruptions", () => {
 
         await getAllDisruptions(req, res);
 
-        expect(getDisruptionsSpy).toHaveBeenCalledOnce();
+        expect(getDisruptionsDataFromDynamoSpy).toHaveBeenCalledOnce();
+        expect(sortDisruptionsByStartDateSpy).toHaveBeenCalledOnce();
 
         expect(res.status).toHaveBeenCalledWith(200);
     });
 
-    it("should return 403 when no session set", async () => {
+    it("should redirect to error page when no session set", async () => {
         getSessionSpy.mockReturnValue(null);
 
         const { req, res } = getMockRequestAndResponse({
@@ -73,17 +143,32 @@ describe("getAllDisruptions", () => {
 
         await getAllDisruptions(req, res);
 
-        expect(getDisruptionsSpy).not.toHaveBeenCalledOnce();
+        expect(getDisruptionsDataFromDynamoSpy).not.toHaveBeenCalledOnce();
+        expect(sortDisruptionsByStartDateSpy).not.toHaveBeenCalledOnce();
 
         expect(res.status).toHaveBeenCalledWith(403);
     });
 
-    it("should return 403 if orgIds do not match", async () => {
-        getDisruptionsSpy.mockResolvedValue("");
+    it("should call filterDisruptionsForOperatorUser for operator users", async () => {
+        getSessionSpy.mockImplementation(() => {
+            return { ...mockSession, isOperatorUser: true, operatorOrgId: "test-operator" };
+        });
+
+        const operatorDisruptions = disruptions.map((disruption) => {
+            return { ...disruption, createdByOperatorOrgId: "test-operator" };
+        });
+
+        const disruptionsWithOperatorDisruptions = [...operatorDisruptions, ...disruptions];
+
+        filterDisruptionsForOperatorUserSpy.mockImplementation(() => {
+            return operatorDisruptions;
+        });
+
+        getDisruptionsDataFromDynamoSpy.mockResolvedValue({ disruptions: disruptionsWithOperatorDisruptions });
 
         const { req, res } = getMockRequestAndResponse({
             query: {
-                organisationId: "123",
+                organisationId: DEFAULT_ORG_ID,
             },
             mockWriteHeadFn: writeHeadMock,
         });
@@ -94,39 +179,88 @@ describe("getAllDisruptions", () => {
 
         await getAllDisruptions(req, res);
 
-        expect(getDisruptionsSpy).not.toHaveBeenCalledOnce();
+        expect(getDisruptionsDataFromDynamoSpy).toHaveBeenCalledOnce();
+        expect(sortDisruptionsByStartDateSpy).toHaveBeenCalledOnce();
 
-        expect(res.status).toHaveBeenCalledWith(403);
+        expect(filterDisruptionsForOperatorUserSpy).toHaveBeenCalledOnce();
+        expect(res.status).toHaveBeenCalledWith(200);
     });
 
-    describe("getDashboardDisruptions", () => {
-        it("should correctly parse disruptions", async () => {
-            getDisruptionsSpy.mockResolvedValue(JSON.stringify(disruptionArray));
+    describe("formatSortedDisruptions", () => {
+        MockDate.set("2023-10-18");
 
-            const disruptions = await getDisruptionsForTable(DEFAULT_ORG_ID, false);
-
-            expect(getDisruptionsSpy).toHaveBeenCalledOnce();
-
-            expect(disruptions).toMatchSnapshot();
+        it("correctly formats disruptions", () => {
+            const formatted = formatSortedDisruption(sortedDisruption);
+            expect(formatted).toMatchSnapshot();
         });
 
-        it("should filter out duplicate disruptions", async () => {
-            const disruptionsToStringify = [
-                ...disruptionArray,
-                {
-                    ...disruptionWithConsequencesAndSocialMediaPosts,
-                    disruptionId: "ca090776-57c6-46a1-a03a-6e0236ee17c8",
-                },
-            ];
-            disruptionsToStringify[2].publishStatus = PublishStatus.pendingAndEditing;
+        it("correctly formats disruptions to live", () => {
+            const formatted = formatSortedDisruption({
+                ...sortedDisruption,
+                validity: [
+                    {
+                        disruptionStartDate: "25/03/2022",
+                        disruptionStartTime: "1123",
+                        disruptionEndDate: "30/03/2022",
+                        disruptionEndTime: "1123",
+                    },
+                    {
+                        disruptionStartDate: "25/12/2040",
+                        disruptionStartTime: "1123",
+                        disruptionEndDate: "30/12/2040",
+                        disruptionEndTime: "1123",
+                    },
+                    {
+                        disruptionStartDate: "25/03/2090",
+                        disruptionStartTime: "1123",
+                        disruptionEndDate: "30/03/2090",
+                        disruptionEndTime: "1123",
+                    },
+                ],
+            });
 
-            getDisruptionsSpy.mockResolvedValue(JSON.stringify(disruptionsToStringify));
+            expect(formatted.isLive).toEqual(true);
+        });
+        it("correctly formats disruptions to upcoming", () => {
+            const formatted = formatSortedDisruption({
+                ...sortedDisruption,
+                validity: [
+                    {
+                        disruptionStartDate: "25/03/2090",
+                        disruptionStartTime: "1123",
+                        disruptionEndDate: "30/03/2090",
+                        disruptionEndTime: "1123",
+                    },
+                    {
+                        disruptionStartDate: "25/12/2090",
+                        disruptionStartTime: "1123",
+                        disruptionEndDate: "30/12/2090",
+                        disruptionEndTime: "1123",
+                    },
+                ],
+            });
+            expect(formatted.isLive).toEqual(false);
+        });
+        it("correctly formats disruptions to recently closed", () => {
+            const formatted = formatSortedDisruption({
+                ...sortedDisruption,
+                validity: [
+                    {
+                        disruptionStartDate: "25/03/2021",
+                        disruptionStartTime: "1123",
+                        disruptionEndDate: "30/03/2021",
+                        disruptionEndTime: "1123",
+                    },
+                    {
+                        disruptionStartDate: "10/10/2023",
+                        disruptionStartTime: "1123",
+                        disruptionEndDate: "11/10/2023",
+                        disruptionEndTime: "1123",
+                    },
+                ],
+            });
 
-            const disruptions = await getDisruptionsForTable(DEFAULT_ORG_ID, false);
-
-            expect(getDisruptionsSpy).toHaveBeenCalledOnce();
-
-            expect(disruptions.length).toBe(3);
+            expect(formatted.isLive).toEqual(false);
         });
     });
 });
