+++ site/pages/disruption-detail/[disruptionId].page.tsx
@@ -24,7 +24,7 @@ import {
     TYPE_OF_CONSEQUENCE_PAGE_PATH,
 } from "../../constants";
 import { getDisruptionById } from "../../data/dynamo";
-import { getS3SignedUrl } from "../../data/s3";
+import { getItem } from "../../data/s3";
 import { ErrorInfo, PageState } from "../../interfaces";
 import { FullDisruption } from "../../schemas/disruption.schema";
 import { SocialMediaPost, SocialMediaPostTransformed } from "../../schemas/social-media.schema";
@@ -924,11 +924,8 @@ export const getServerSideProps = async (
             disruption.socialMediaPosts.map(async (s) => {
                 if (s.image) {
                     const url: string =
-                        (await getS3SignedUrl(
-                            process.env.IMAGE_BUCKET_NAME || "",
-                            s.image?.key,
-                            s.image?.originalFilename,
-                        )) || "";
+                        (await getItem(process.env.IMAGE_BUCKET_NAME || "", s.image?.key, s.image?.originalFilename)) ||
+                        "";
                     return {
                         ...s,
                         image: {
